pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "un-algebra-reanchor"
version = "0.1.0"
description = "UN-Algebra retro-validation tests for metrology datasets"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = {text = "MIT"}
authors = [{name="Eric D. Martin", email="eric@aybllc.org"}]
dependencies = [
  "pandas==2.2.2",
  "numpy==1.26.4",
  "PyYAML==6.0.1"
]

[project.scripts]
unreanchor = "un_reanchor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["un_reanchor"]

[tool.pytest.ini_options]
addopts = "-q"
